Impl userGold.confirmedAt data strcture
---

Original commit (chain-contracts): fcf38743a476f4092fadf4c98cb933756b563eb1

+++ chain-contracts/contracts/Logic.sol
@@ -37,7 +37,7 @@ contract Logic {
     if (userNormalPlanet.balanceOf(msg.sender) == 0 && gold.balanceOf(
       msg.sender
     ) == 0) {
-      gold.mint(msg.sender, SafeMath.sub(10, planetPrice));
+      gold.mint(msg.sender, uint200(SafeMath.sub(10, planetPrice)));
     } else {
       gold.unmint(msg.sender, planetPrice);
     }
